KUDU-3090: Native owner metadata in Kudu
Apache Sentry and Apache Ranger both support permissions granted to
table owners, but as Sentry integrates with Apache Hive Metastore (HMS)
and stores its metadata in it, Kudu didn't need to store table ownership
to support granting permissions to owners.

Apache Ranger on the other hand doesn't depend on HMS and needs Kudu to
tell it if the owner is attempting to authorize an action, so to enable
users to grant privileges to owners we need to support ownership
natively.

This patch adds the basic plumbing for table ownership, synchronizing
ownership metadata with HMS both using the notification log listener and
via tooling, and setting the owner on CREATE TABLE and ALTER TABLE
requests in the C++ client.

The maximum owner length is 128 characters by default which aligns with
HMS/Apache Impala maximum owner lengths, but it's configurable with the
max_owner_length flag.

Supporting this in the Java and Python clients, authorizing these
requests, and support for ownership in authorization will come in
follow-up patches.

Credit goes to Grant Henke <granthenke@apache.org> for the initial
version of this patch.

Design doc: https://s.apache.org/kudu-ownership-design

Change-Id: I67f5bfdf56d409960365fd5803913a2d3800831d
Reviewed-on: http://gerrit.cloudera.org:8080/15841
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>