KUDU-3154: add policy to Ranger before authorizing
For whatever reason, in some environments, we've seen the Kudu Ranger
plugin hang when downloading policies. It isn't exactly clear why this
happens but one workaround we've seen is to add a policy to Ranger
before authorizing any requests. This patch adds the workaround to the
only Ranger test that suffers from this.

Change-Id: Ice0b3842335e5854835542e18b41c54509c2fc33
Reviewed-on: http://gerrit.cloudera.org:8080/16166
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Attila Bukor <abukor@apache.org>