We return the pending config via Raft Consensus if it exists
Summary: In order to return a pair of committed as well
as pending config to automation/clients, we need to access
the pending config. We add a simple function for this.

Test Plan: We wrote an end 2 end test on facebook code side,
which creates a small ring and accesses the Configuration state
pair during the config change event

Reviewers: bhatvinay, iritwik

Subscribers:

Tasks:

Tags: