[ranger] Add refreshing Ranger policies
Authorization tests depend on frequent polling of policies from the
Ranger server to make sure newly created policies are cached in the
authorizer plugin. We've seen some flaky tests due to this recently, so
a better way to handle this would be to manually refresh the policies in
the authorizer plugin.

When the Ranger integration was initially implemented we decided to not
use the existing ResetAuthzCache methods in master as it behaves in a
different way than it did with Sentry. Sentry's cache was within Kudu
and could be fully reset. Unfortunately Ranger's cache lives in the
Ranger plugin within the Ranger subprocess which has a void
refreshPolicies() method that doesn't throw any checked exceptions. This
means there's no way to guarantee the cache is cleared and the only
failure we can know about is if the request times out or other failures
between the master and the subprocess, but not between the subprocess
and the Ranger server. This means that even if Ranger itself is down, we
can still get an OK response when calling this method and there's no way
to guarantee clearing the cache.

When Sentry integration was removed from Kudu, this ResetAuthzCache()
method was also removed along with the tooling, as we didn't have any
authz providers that would support resetting the cache.

Now we need a way to do this to deflake the tests, and as there is no
longer a discrepancy between the behavior of different authz providers
and the best effort basis of refreshing policies should be enough, this
method is readded by this commit and the tests are changed to use it
instead of the timer-based policy refreshes.

Change-Id: Ie90b1c23b92060e081ab514fbae417b4f31c2b66
Reviewed-on: http://gerrit.cloudera.org:8080/16160
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>