loadgen: use multiple KuduClient instances
KuduClient is threadsafe but doesn't perform well under high
concurrency. This adds a flag (defaulted true) to 'kudu perf loadgen'
that lets it use a separate client for each inserter thread. This speeds
up the insertion substantially (about 30% in my quick local test, but
could be more depending on the machine and target cluster specs)

Change-Id: I722a0761e97894857e2ca3c0286b4467d7f27d47
Reviewed-on: http://gerrit.cloudera.org:8080/16184
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>