KUDU-2612 p7: add transaction participants to tablets
This patch adds state tracking on transaction participants (i.e.
tablets that are participating in a transaction) in the form of a map
from transaction ID to transaction state.

Each transaction that a tablet participates in has associated with it a
state machine that will be controlled by the transaction status manager
based on global transaction state maintained therein. Participant state
tracking is plugged into the op driver framework. The details of this
state machine are outlined in txn_participant.h.

While this patch Raft commits participant ops using the existing op
driver framework, it doesn't replay participant ops from the WAL or
anchor WAL segments. As such, tests are only added to test the
replication aspect of participant op drivers, rather than testing any
form of durability.

The added state can only be called via internal APIs. There are no RPC
endpoints with which the code in this patch can be reached.

Change-Id: I39201ce1d911308cd28f3f4790a126e30052f3bf
Reviewed-on: http://gerrit.cloudera.org:8080/16277
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>