Allow for first run log init recovery if application says so
Summary: This is for the case where the logs are in a different
directory that consensus-metadata. The init of the logs helps
figure out the Last Log Opid and Committed Opid, while the
consensus metadata is for the term and the committed config, which
is still missing. In the future, we plan to expect a copied
consensus-metadata from a previous instance.

The workaround to allow setting the term unblocks witness mode
for now.

Test Plan: used this to check that a witness does not start
from 1:1

Reviewers: yashtc

Subscribers:

Tasks:

Tags: