KUDU-3170: deflake testSlowWriterDoesntBlockQueues
Previously the test would block the writer thread for 1s, expect a
timeout to happen within 1s, and check that the slow writer still
allowed for the outbound queue to be inserted to.

This led to some flakiness, as the first write may have sometimes
completed within 1s and allowed the second request to leave the outbound
queue and begin writing. This patch addresses this by extending the time
with which we block writes to extend past the execution deadline.

Without this patch, we saw frequent failures in ARM runs of the test,
and I saw 4/400 failures in dist-test. With the patch, the test passed
in dist-test 1600/1600 times.

Change-Id: I8b322782630b3ef935fb0264b6195d2ed72cc58c
Reviewed-on: http://gerrit.cloudera.org:8080/16221
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Andrew Wong <awong@cloudera.com>