In flexi-raft mode we pick up local-peer from server config
Summary: As local peer was not setting flexi-raft we were failing
a CHECK in consensus_queue.cc. In flexi-raft mode we ensure
that regions and backed_by_database is set.

Test Plan: Tested this in local setups with bootstrap. Rings
are now properly bootstrapping in flexi-raft mode and not aborting
with the CHECK

Reviewers: iRitwik

Subscribers:

Tasks:

Tags: