KUDU-3090 Allow restoring tables without owner
fad779cf7 introduced backing up and restoring table ownership as part of
the backup tool. 6f0e858ba restricted changing ownership to users with
ALL + delegate admin privileges which is higher than what we required
for running restore. To support running the restore job without having
to grant additional privileges, this commit adds a new "restoreOwner"
flag to the restore job which is true by default, but if set to false,
the job creates tables as the logged in user which still requires the
same privileges as before.

Change-Id: I22d0b78858f60fafd3c116c7e4b3371839440d6c
Reviewed-on: http://gerrit.cloudera.org:8080/16239
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>