KUDU-1802: Add KuduScanner.GetKuduTable
This patch adds the ability to get the KuduTable instance from the
KuduScanner. This allows the user to use the KuduTable instance
populated by the scan token instead of making a GetTableSchema
call to the master.

The complete KuduTable is passed to the KuduScanner in the
constructor, this patch allows it to be re-used by the consumer of the
scanner. It's literally a passthrough, but it helps when the KuduTable
was constructed internally (e.g via a ScanToken).

Change-Id: I09a80c6d499987553aef1338db93397a1de2491e
Reviewed-on: http://gerrit.cloudera.org:8080/16251
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>