KUDU-2612 p9: anchor participant ops in WAL
Participant ops will now anchor WAL segments in the same way that we
anchor WAL segments for other in-memory state (e.g. MRS, DMS) that gets
rebuilt upon recovery.

Each participant op for a given transaction will update the single
anchor associated with that transaction. As a transaction is
transitioned from state to state, the transaction's prior participant op
is unanchored in favor of anchoring the new op.

Since there currently isn't a need to rid a participant of in-memory
state associated with a transaction, this adds a test-only method to
remove a committed or aborted transaction -- once this method called, a
transaction's anchor is removed and WAL GC may proceed to remove the
latest participant op.

Change-Id: I936f0a345c4b6095f0d99b6dd244e3092ae3f9d7
Reviewed-on: http://gerrit.cloudera.org:8080/16358
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>