KUDU-1587 part 2: reject write ops if apply queue is overloaded
This patch implements control admission for write requests in tablet
servers based on the load status of their apply queue. With this change,
the recently introduced OpApplyQueueTest.ApplyQueueBackpressure scenario
successfully passes.

If the queue times of the tasks in the apply queue become higher than
the specified threshold, the apply queue enters overloaded state.  When
the queue is overloaded, the tablet server rejects incoming write
requests with some probability.  The longer the queue stays overloaded,
the greater the probability of rejections.  The apply queue exits the
overloaded state when queue times drop below the specified threshold.

This new behavior is not yet enabled by default, keeping the legacy
behavior of unbounded/uncontrolled queue times as is.  To enable it,
set --tablet_apply_pool_overload_threshold_ms to something greater
than 0 (e.g., 500).

Change-Id: I6d7688d6fa832e606b8efc4549568fa52dfa1931
Reviewed-on: http://gerrit.cloudera.org:8080/16343
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>