Upgrade docs/site building tools
The tooling to build the documentation for the Kudu website are very
old making it difficult to run it on newer OSes.

This patch upgrades the Ruby dependencies along with the required
Doxygen version to make it easier. As the new dependencies are fairly
new, they require Ruby and Doxygen versions newer than what's included
by default on CentOS 7 and likely on other OSes that we might want to
use to build the site. It's fairly simple to upgrade them on the other
hand, this patch also adds instructions to README.adoc.

Change-Id: I5eddf689da90123ca52075d09de07981d11f8ffe
Reviewed-on: http://gerrit.cloudera.org:8080/16371
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>