KUDU-3187: Enhance the HMS plugin to check if synchronization is enabled
This patch enhances the HMS plugin to check if synchronization
is enabled on the Kudu cluster backing the table. This can
simplify enabling HMS synchronization because the plugin
jar can be added to the HMS classpath unconditionally.

This also helps support cases where multiple Kudu clusters
use the same HMS and one is synchronized and the other isn’t.

This patch introduces a new environment property,
`KUDU_HMS_SYNC_ENABLED`, to the plugin which allows
for faster tests that do not need to setup a Kudu cluster.
It could also be useful to disable this new functionality as
a workaround if issues are seen.

Additonally, this patch adjusts the mini HMS to add
security properties configuration. This is required now
that the Kudu client runs in the plugin and communicates
with the cluster. See these commits for context:
- https://github.com/apache/kudu/commit/0ee93e31a1febb987c72e7392a69b2584e6f38ed
- https://github.com/apache/kudu/commit/3343144fefaad5a30e95e21297c64c78e308fa1f

I also manually verified the following behavior on a real cluster:
- `KUDU_HMS_SYNC_ENABLED` environment variable
- CREATE/ALTER/DROP on non-sync cluster works
- - With plugin jar used by the HMS
- CREATE/ALTER/DROP on sync cluster fails with clear message
- Authn via Kerberos

Change-Id: Ib3588d72af1bb499202b47fca50a08876e13ea37
Reviewed-on: http://gerrit.cloudera.org:8080/16388
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>