KUDU-2574: Add a unique cluster ID
This patch adds a unique cluster ID to the cluster. The ID is
a UUID that is automatically generated and stored in the
sys_catalog if missing (on fresh startup or upgrade).

This patch also exposes the cluster ID via the masters web-ui
and the `kudu master list` tool.

I plan to use this cluster ID to enhance the HMS sync
functionality, but it can be exposed further as
needed for other uses as well.

Change-Id: I4df1d3f7d100336f52563f3008cacf6d9e328fae
Reviewed-on: http://gerrit.cloudera.org:8080/16403
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins