Fix a Google+ link
When fixing KUDU-2260 a Google+ link was added for additional context in
a comment. As Google+ was shut down since then, this link is now dead.
Tried replacing it with a Wayback Machine link, but it doesn't seem to
load anymore either. Unfortunately I couldn't find a more formal
documentation for this, but at least Alexey managed to find an
explanation on Hacker News (there also happens to be a link to the
Google+ post in that thread). Hopefully it stays online for far longer
than Google+.

Change-Id: I7e10e49b08428079d3134e9291328e4508822ecf
Reviewed-on: http://gerrit.cloudera.org:8080/16279
Tested-by: Attila Bukor <abukor@apache.org>
Reviewed-by: Grant Henke <granthenke@apache.org>