KUDU-2884: [hms] Improve master address matching
This patch improves the master address matching in the
`kudu hms fix` tool. Instead of expecting a literal match
of the address strings, it expects the match to be functionally
the same. This means that the addresses can be missing the
port (using the default), duplicated, and re-ordered. The
HMS will still consider them the same if they match the
clusters masters functionally.

Though the default port is considered when matching master
addresses, it is not included in the automated hms tool tests
given the tests use random ephemeral ports. Instead a test
for MasterAddressesToSet and UnorderedHostPortSet equality
was added.

Change-Id: I00059ff1b5134f7e5d6ea8be4fdf0701aa56b0c7
Reviewed-on: http://gerrit.cloudera.org:8080/16420
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins