KUDU-3157: Ensure we don’t shade slf4j classes
We should almost never shade slf4j classes but due to a behavior
issue in the Shadow Gradle plugin, exclusions from parent
modules are not respected in modules that use them. This meant
that we were shading slf4j classes into the kudu-spark module
jars and any modules that depend on kudu-spark.

This patch fixes the issue by universally excluding slf4j from
being shaded unless `shadowIncludeSlf4j` is explicitly specified.
Currently we only do this in the kudu-subrocess module.

Change-Id: I72605594bdc70670f82eff4154247c5c0252de7f
Reviewed-on: http://gerrit.cloudera.org:8080/16453
Tested-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Grant Henke <granthenke@apache.org>