test: deflake tablet_server_quiescing-itest
It was possible that TestScansRetry would fail with more than the one
active scanner at the end of the test. This is possible because the
client scanner in the test that may have lingering retries from a
previous scan attmpts.

This patch updates to allow for multiple scan attempts, since it doesn't
detract from testing the fact that the keep-alive calls are working.

It also reenstates the quiescing state before calling the keep-alives,
which appears to be the original intent based on comments.

Change-Id: Ib6525b913cbe9acbdbc59d8a52df81c85327d0c7
Reviewed-on: http://gerrit.cloudera.org:8080/16463
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins