test: reduce the number of files used by TestDuplicatedRowsRandomCompaction
The test generates a large number (10 + 9 + ... + 1 = 55) of rowsets
with a schema containing 3 columns and a workload that has creates delta
files. On MacOS this is problematic, as 1) by using the file block
manager, we create a single file per CFile, and 2) the default file
limit is 256 (at least on macOS 10.15.4).

This patch alleviates this by adjusting the test to use fewer (5 + 4 +
... 1 = 15) rowsets. Despite the randomness in the workload, the test
passed on my machine 100/100 times.

Change-Id: Idf938f24b4a79537337f7a53d731f3162716145d
Reviewed-on: http://gerrit.cloudera.org:8080/16464
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Grant Henke <granthenke@apache.org>