tablet: add comment to summarize the bootstrap
In the few times I've had to peruse tablet bootstrapping code, I've been
disappointed there isn't a high-level summary of what the bootstrap
process entails with regards to WAL replay. So, I took a stab at it in
hopes it might make the job for future readers a bit easier.

Change-Id: I32d699efd97a3fdf57f47c30eded53f93f4553b7
Reviewed-on: http://gerrit.cloudera.org:8080/16468
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins