[kuduraft] change read api into log to have the uuid of the destination
Summary:
Adds the uuid parameter into the read api of log_cache and log:
LogCache::ReadOps() and Log::ReadReplicatesInRange(). This is needed so that the
BinlogWrapper can have access to the destination uuid to reprt error when the
OpId to be read is not found either in th index chunk or in the binlog file.

Test Plan:

Reviewers: arahut

Subscribers:

Tasks:

Tags: